+++ packages/asc-ui/src/components/Button/ToggleButton/ToggleButtonStyle.ts
@@ -1,5 +1,5 @@
 import styled from '@datapunt/asc-core'
-import { color } from '../../../utils'
+import { themeColor } from '../../../utils'
 import ButtonStyle, { Props as ButtonProps } from '../ButtonStyle'
 
 export type Props = {
@@ -9,11 +9,11 @@ export type Props = {
 
 export default styled(ButtonStyle)<Props>`
   background-color: ${({ open, theme }) =>
-    color('tint', open ? 'level2' : 'level1')({ theme })};
+    themeColor('tint', open ? 'level2' : 'level1')({ theme })};
 
   &:hover,
   &:focus {
     background-color: ${({ theme, open }) =>
-      color('tint', open ? 'level3' : 'level2')({ theme })};
+      themeColor('tint', open ? 'level3' : 'level2')({ theme })};
   }
 `