+++ packages/asc-ui/src/components/Card/CardStyle.ts
@@ -1,5 +1,5 @@
 import styled, { Theme, css } from '@datapunt/asc-core'
-import { color } from '../../utils'
+import { themeColor } from '../../utils'
 import { perceivedLoading } from '../../utils/themeUtils'
 import CardActionsStyle from './CardActionsStyle'
 import { CardMediaWrapperStyle } from './CardMediaStyle'
@@ -16,9 +16,9 @@ export type Props = {
 const CardStyle = styled.div<Props>`
   ${({ maxWidth }) => maxWidth && `max-width: ${maxWidth}px;`}
   ${({ shadow, theme }) =>
-    shadow && `box-shadow: 2px 2px ${color('tint', 'level3')({ theme })};`}
+    shadow && `box-shadow: 2px 2px ${themeColor('tint', 'level3')({ theme })};`}
   background-color: ${({ backgroundColor, theme }) =>
-    color('tint', backgroundColor || 'level1')({ theme })};
+    themeColor('tint', backgroundColor || 'level1')({ theme })};
   display: flex;
   flex-direction: ${({ horizontal }) => (horizontal ? 'row' : 'column')};
   justify-content: space-between;