+++ packages/asc-ui/src/components/Editorial/EditorialBody/EditorialBodyStyle.ts
@@ -1,9 +1,9 @@
 import styled from '@datapunt/asc-core'
-import { color } from '../../../utils'
+import { themeColor } from '../../../utils'
 
 export type Props = {}
 
 export default styled.article<Props>`
   flex: 1;
-  background-color: ${color('tint', 'level1')};
+  background-color: ${themeColor('tint', 'level1')};
 `