+++ packages/asc-ui/src/components/Footer/FooterBottom/FooterBottomStyle.ts
@@ -1,8 +1,8 @@
 import styled from '@datapunt/asc-core'
-import { color } from '../../../utils'
+import { themeColor } from '../../../utils'
 
 export type Props = {}
 
 export default styled.div<Props>`
-  background-color: ${color('tint', 'level1')};
+  background-color: ${themeColor('tint', 'level1')};
 `