+++ packages/asc-ui/src/components/Menu/MenuFlyOut/MenuFlyOutStyle.ts
@@ -2,7 +2,7 @@ import styled, { css } from '@datapunt/asc-core'
 import { MenuItemStyle } from '../MenuItem'
 import MenuListStyle from '../MenuList/MenuListStyle'
 import { FLY_OUT_WIDTH, MENU_ITEM_SIZE } from '../constants'
-import { color } from '../../../utils'
+import { themeColor } from '../../../utils'
 
 export type Props = {
   hasToggle?: boolean
@@ -10,7 +10,7 @@ export type Props = {
 
 export default styled(MenuItemStyle)<Props>`
   ${MenuListStyle} {
-    background-color: ${color('tint', 'level2')};
+    background-color: ${themeColor('tint', 'level2')};
     width: ${FLY_OUT_WIDTH}px;
     top: ${MENU_ITEM_SIZE}px;
     order: 0;
@@ -20,7 +20,7 @@ export default styled(MenuItemStyle)<Props>`
     ${({ hasToggle }) =>
       !hasToggle &&
       css`
-        border-bottom: 4px solid ${color('tint', 'level7')};
+        border-bottom: 4px solid ${themeColor('tint', 'level7')};
       `}
 
     & > ${MenuItemStyle} {