+++ packages/asc-ui/src/components/Tag/Tag.test.tsx
@@ -1,7 +1,7 @@
 import { ascDefaultTheme, ThemeProvider } from '@datapunt/asc-core'
 import { cleanup, render } from '@testing-library/react'
 import * as React from 'react'
-import { color } from '../../utils'
+import { themeColor } from '../../utils'
 import Tag from './Tag'
 
 const theme = {
@@ -20,22 +20,22 @@ describe('Tag', () => {
     expect(queryByText(/Tag one/)).not.toBeNull()
     expect(container.firstChild).toHaveStyleRule(
       'background-color',
-      color('tint', 'level2')({ theme }),
+      themeColor('primary', 'main')({ theme }),
     )
     // color should be readable
-    expect(container.firstChild).toHaveStyleRule('color', '#000')
+    expect(container.firstChild).toHaveStyleRule('color', '#fff')
   })
   it('should render and apply custom style', () => {
     const { container, queryByText } = render(
       <ThemeProvider theme={theme}>
-        <Tag colorVariant="primary">Tag two</Tag>
+        <Tag colorType="primary">Tag two</Tag>
       </ThemeProvider>,
     )
 
     expect(queryByText(/Tag two/)).not.toBeNull()
     expect(container.firstChild).toHaveStyleRule(
       'background-color',
-      color('primary')({ theme }),
+      themeColor('primary')({ theme }),
     )
     // color should be readable
     expect(container.firstChild).toHaveStyleRule('color', '#fff')