+++ packages/asc-ui/src/utils/index.ts
@@ -1,7 +1,6 @@
 export { default as fromTheme, valueFromObject } from './fromTheme'
 export {
-  color,
-  getColorCode,
+  themeColor,
   getTypographyFromTheme,
   focusStyleOutline,
   srOnlyStyle,